requirements.txt
streamlit>=1.28.0
st-supabase-connection>=1.0.0
streamlit-code-editor>=0.1.0
python-dotenv>=1.0.0